import: check for memory allocation issues
Builds on gcc10 quite frequently fail due to memory allocation issues
that can be detected by checking for the message "fork: Cannot
allocate memory". This commit detects the builds affected by this
problem and rejects them.

+++ web/import.inc.php
@@ -45,6 +45,13 @@ function import_result($buildid, $filename)
       return;
     }
 
+    system("grep -q 'fork: Cannot allocate memory' " . $thisbuildtmpdir . "build-end.log",
+	   $retval);
+    if ($retval == 0) {
+      echo "Reject build result, there was a memory allocation problem\n";
+      return;
+    }
+
     /* Remove the build.log.bz2 file if it's in there */
     system("rm -f " . $thisbuildtmpdir . "build.log.bz2", $retval);
